fix: fix tasks suites
