+++ src/resources/boards/column.model.ts
@@ -2,7 +2,7 @@
 const { v4: uuidv4 } = require('uuid');
 
 export interface IColumn {
-  id:string;
+  id?:string;
   title:string;
   order:number;
 }
@@ -18,7 +18,7 @@ class Column {
       id = uuidv4(),
       title = "Djdsdddsd",
       order = 0,
-    } : IColumn) {
+    } = {}) {
       this.id = id;
       this.title = title;
       this.order = order;
